Move Wait its own file.
If anything, this will make testing and the change more localized.
